spark/requirements.txt
spark
numpy
sklearn
pyspark
scipy
kafka-python==1.3.1
